Add new metrics

+++ API/UserCustomIndices/Model/Response/TrailingPerformance.cs
@@ -3,7 +3,7 @@
     public struct TrailingPerformance
     {
         public int TimePeriod;
-        public decimal Upper;
-        public decimal Lower;
+        public double Upper;
+        public double Lower;
     }
 }
